Fold offeranswer.js into caress.js
The separation made sense back in early designs, when offeranswer.js was "simple".
Now, however, offeranswer's routes make up 90% of the app, and the routes are
available in separate files, so it's not a big deal to put them all under one
roof (makes some of the middleware assumptions a little less janky, too).