Remove text/* content-type requirement for body parsing
Really, no matter what the type read is, this app is going to treat the content
the same, so we should just read everything as-is. More importantly: we should
be ready for arbitrary binary data, and as such should not convert the body from
a Buffer to text before handling (as we treat it like a blob anyway).

At some point soon (possibly preceding the Lua switch), the Redis
connection should be switched to using the {return_buffers: true}
option, with explicit calls to .toString in the (surprisingly few)
cases where that is necessary.