Move scripts into own newline-separated modules
The newline separation is a big advantage: no more fleeting
semicolon/whitespace issues, the only syntax that matters is mirrored
in the JS, and we get line numbers if things break.

Also, I think this fixes the failures in previous versions (one of the
pseudo-lines was missing a space at the end).